+++ hummingbird/commons/src/main/java/com/logica/hummingbird/spacesystemmodel/parameters/behaviours/LongUnsignedBehaviour.java
@@ -12,7 +12,7 @@ import com.logica.hummingbird.spacesystemmodel.exceptions.InvalidParameterTypeEx
  */
 public class LongUnsignedBehaviour extends AbstractIntegerBehaviour {
 	
-	public LongUnsignedBehaviour(long sizeInBits) throws InvalidParameterTypeException {
+	public LongUnsignedBehaviour(int sizeInBits) throws InvalidParameterTypeException {
 		super(sizeInBits);
 		if(sizeInBits > 64) {
 			throw new InvalidParameterTypeException("LongUnsigned cannot be greater than 64-bits in size.");
